requirements.txt
Flask
flask-mysqldb==0.2.0
Pillow